Refactor: move service validations to Service#validate
