+++ lib/Serverless.js
@@ -66,16 +66,6 @@ class Serverless {
 
     return this.service.load(this.processedInput.options)
       .then(() => {
-        this.service.listDuplicateHandlers().forEach(arg => {
-          const handlerName = arg[0];
-          const group = arg[1];
-          const warnMessage = [
-            `A handler "${handlerName}" is duplicated in functions: `,
-            `${group.join(', ')}.`,
-          ].join('');
-          this.cli.log(warnMessage);
-        });
-
         // load all plugins
         this.pluginManager.loadAllPlugins(this.service.plugins);
