+++ lib/Serverless.test.js
@@ -179,41 +179,6 @@ describe('Serverless', () => {
         expect(updateAutocompleteCacheFileStub.calledOnce).to.equal(true);
       });
     });
-
-    it('should warn if multiple functions have same handler', () => {
-      class CustomCLI extends CLI {}
-      serverless.classes.CLI = CustomCLI;
-      const consoleLogStub = sinon.stub(CustomCLI.prototype, 'log');
-
-      const SUtils = new Utils();
-      const tmpDirPath = testUtils.getTmpDirPath();
-      const serverlessYml = {
-        service: 'testService',
-        provider: 'testProvider',
-        plugins: ['testPlugin'],
-        functions: {
-          functionA: {
-            handler: 'foo.functionA',
-          },
-          functionB: {
-            handler: 'foo.functionA',
-          },
-          function123: {
-            handler: 'foo.function123',
-          },
-          function456: {
-            handler: 'foo.function123',
-          },
-        },
-      };
-      SUtils.writeFileSync(path.join(tmpDirPath, 'serverless.yml'),
-        YAML.dump(serverlessYml));
-      serverless.config.update({ servicePath: tmpDirPath });
-
-      return serverless.init().then(() => {
-        expect(consoleLogStub.callCount).to.equal(2);
-      });
-    });
   });
 
   describe('#run()', () => {
