+++ lib/classes/Service.js
@@ -208,10 +208,20 @@ class Service {
       }
     });
 
+    this._listDuplicateHandlers().forEach(arg => {
+      const handlerName = arg[0];
+      const group = arg[1];
+      const warnMessage = [
+        `A handler "${handlerName}" is duplicated in functions: `,
+        `${group.join(', ')}.`,
+      ].join('');
+      this.serverless.cli.log(warnMessage);
+    });
+
     return this;
   }
 
-  listDuplicateHandlers() {
+  _listDuplicateHandlers() {
     const functionObjs = this.getAllFunctions()
       .map(name => ({
         name,
