+++ lib/classes/Service.test.js
@@ -9,6 +9,7 @@ const Service = require('../../lib/classes/Service');
 const Utils = require('../../lib/classes/Utils');
 const Serverless = require('../../lib/Serverless');
 const testUtils = require('../../tests/utils');
+const CLI = require('../../lib/classes/CLI');
 
 // Configure chai
 chai.use(require('chai-as-promised'));
@@ -720,6 +721,46 @@ describe('Service', () => {
         // validate the service configuration, now that variables are loaded
         expect(() => serverless.service.validate())
           .to.throw('Events for "functionA" must be an array, not an string');
+      });
+    });
+
+    it('should warn if multiple functions have same handler', () => {
+      const SUtils = new Utils();
+      const serverlessYml = {
+        service: 'testService',
+        provider: 'testProvider',
+        functions: {
+          functionA: {
+            handler: 'foo.functionA',
+            events: [],
+          },
+          functionB: {
+            handler: 'foo.functionA',
+            events: [],
+          },
+          function123: {
+            handler: 'foo.function123',
+            events: [],
+          },
+          function456: {
+            handler: 'foo.function123',
+            events: [],
+          },
+        },
+      };
+      SUtils.writeFileSync(path.join(tmpDirPath, 'serverless.yml'),
+        YAML.dump(serverlessYml));
+
+      const serverless = new Serverless({ servicePath: tmpDirPath });
+
+      class CustomCLI extends CLI {}
+      const consoleLogStub = sinon.stub(CustomCLI.prototype, 'log');
+      serverless.cli = new CustomCLI(serverless);
+
+      return expect(serverless.service.load()).to.eventually.be.fulfilled
+      .then(() => {
+        serverless.service.validate();
+        expect(consoleLogStub.callCount).to.equal(2);
       });
     });
   });
